Move basis projection into `*_backward` functions
Specifically, the output of `src_mean_backward` and
`src_covar_backward` now also perform the `evaluate_t` step using
a supplied basis. Previously, this was done in the `conj_grad_*`
functions. It is more appropriate to move it out, however, as this
simplifies tasks such as shrinkage, which would otherwise have to
be included in `conj_grad_covar`.