Fix bug in `conj_grad_vols_wt`
If `conj_grad` gets a matrix as its right-hand side, it assumes that each
column of that matrix is a separate right-hand side and will optimize it
accordingly. If we want it to treat the whole thing as a single right-hand
side, we therefore need to transform it into a single column vector. If we
don't, it will have trouble calculating the norm of the residual and so will
not terminate properly.